pyproject.toml
[project]
name = "datalumina-project-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.1.1",
    "psycopg2-binary>=2.9",
]

[dependency-groups]
dev = [
    "ipykernel>=7.0.0",
]
